+++ example/AsyncLoadHOC/index.js
@@ -1,0 +1,3 @@
+import AsyncLoadHOC from "./AsyncLoadHOC";
+
+export default AsyncLoadHOC;