+++ src/hocs/ArenaSceneLoadMotion/actions.js
@@ -1,9 +1,9 @@
 import { ARENA_SCENE_SET_STATE } from "../../actionTypes";
 import {
   ARENA_SCENE_ANIMATION_NEXTPHRASE,
-  ARENA_SCENE_ANIMATION_LEAVING_START
+  ARENA_SCENE_ANIMATION_LEAVING_START,
+  ARENA_SCENE_ANIMATION_LOAD_BUNDLE
 } from "./actionTypes";
-
 export function setState(state) {
   return {
     type: ARENA_SCENE_SET_STATE,
@@ -21,5 +21,12 @@ export function nextPhase(phase) {
 export function startLeaving() {
   return {
     type: ARENA_SCENE_ANIMATION_LEAVING_START
+  };
+}
+
+export function loadSceneBundle(asyncBundleThunk) {
+  return {
+    type: ARENA_SCENE_ANIMATION_LOAD_BUNDLE,
+    asyncBundleThunk
   };
 }