+++ src/hocs/ArenaSceneLoadMotion/index.js
@@ -1,19 +1,20 @@
 import { bundleToComponent } from "../../tools";
-import ArenaSceneMotion from "./ArenaSceneMotion";
+import ArenaSceneLoadMotion from "./ArenaSceneLoadMotion";
 import * as actions from "./actions";
+import saga from "./saga";
 import reducer from "./reducer";
 
 export default bundleToComponent({
-  Component: ArenaSceneMotion,
+  Component: ArenaSceneLoadMotion,
   actions,
   reducer,
+  saga,
   propsPicker: (state, actions, allState, { _arenaScene }) => ({
     ...state,
     actions,
     reducerKey: _arenaScene.reducerKey
   }),
   options: {
-    vReducerKey: "_arenaSceneAnimation",
-    isSceneReducer: false
+    vReducerKey: "_arenaSceneAnimation"
   }
 });