+++ src/hocs/index.js
@@ -1,6 +1,6 @@
 import ArenaScene from "./ArenaScene";
 import SceneBundle from "./SceneBundle";
-import ArenaSceneMotion from "./ArenaSceneMotion";
+import ArenaSceneLoadMotion from "./ArenaSceneLoadMotion";
 import ReducerDictOverrider from "./ReducerDictOverrider";
 
-export { ArenaScene, SceneBundle, ArenaSceneMotion, ReducerDictOverrider };
+export { ArenaScene, SceneBundle, ArenaSceneLoadMotion, ReducerDictOverrider };