+++ src/index.js
@@ -1,5 +1,5 @@
 export { createArenaStore } from "./core";
-export { ArenaScene, ArenaSceneMotion, ReducerDictOverrider } from "./hocs";
-import * as sceneMotionPhase from "./hocs/ArenaSceneMotion/animationPhase";
+export { ArenaScene, ArenaSceneLoadMotion, ReducerDictOverrider } from "./hocs";
+import * as loadMotionPhase from "./hocs/ArenaSceneLoadMotion/animationPhase";
 
-export { sceneMotionPhase };
+export { loadMotionPhase };