+++ src/tools/index.js
@@ -1,11 +1,4 @@
 import bundleToComponent from "./bundleToComponent";
-import asyncBundleToComponent from "./asyncBundleToComponent";
 import bundleToElement from "./bundleToElement";
-import asyncBundleToElement from "./asyncBundleToElement";
 
-export {
-  bundleToComponent,
-  asyncBundleToComponent,
-  bundleToElement,
-  asyncBundleToElement
-};
+export { bundleToComponent, bundleToElement };