+++ src/utils/replaceReducer.js
@@ -1,8 +1,13 @@
 import {
   ARENA_SCENE_REPLACE_STATE,
+  ARENA_GLOBAL_PROPSPICKER_LOCK
 } from "../core/actionTypes";
 
 export function sceneReplaceReducer(store, reducerKey, reducerFactory, state) {
+  store.dispatch({
+    type: ARENA_GLOBAL_PROPSPICKER_LOCK,
+    lock: true
+  });
   let newReducerKey = store.replaceReducer({
     reducerKey,
     reducer: reducerFactory(reducerKey),
@@ -14,5 +19,9 @@ export function sceneReplaceReducer(store, reducerKey, reducerFactory, state) {
       _sceneReducerKey: newReducerKey,
       state
     });
+  store.dispatch({
+    type: ARENA_PROPSPICKER_UNLOCK,
+    lock: false
+  });
   return reducerKey;
 }