+++ test/integration/ArenaSceneMotion/index.js
@@ -2,9 +2,7 @@ import React from "react";
 import { expect } from "chai";
 import { spy } from "sinon";
 import { createMount } from "../../testUtils";
-import { createArenaStore, sceneMotionPhase } from "src";
-import { bundleToElement } from "src/tools";
-import sceneBundleForTestA from "../../sceneBundleForTestA";
+import { createArenaStore, loadMotionPhase } from "src";
 import TestHOC from "./TestHOC";
 
 function selectAnimationState(allStates, name) {
@@ -32,18 +30,15 @@ describe("<ArenaSceneMotion /> integration", () => {
 
   it("should step into IN phase correctly", () => {
     wrapper = mount(
-      <TestHOC store={store}>
-        {bundleToElement(
-          Object.assign({}, sceneBundleForTestA, {
-            options: { reducerKey: "bundle" }
-          })
-        )}
-      </TestHOC>
+      <TestHOC
+        store={store}
+        asyncBundleThunk={() => import("../../sceneBundleForTestA")}
+      />
     );
     let flagPromise = new Promise(resolve => {
       let unsubscribe = store.subscribe(() => {
         let animationState = selectAnimationState(store.getState());
-        if (animationState && animationState.phase === sceneMotionPhase.IN) {
+        if (animationState && animationState.phase === loadMotionPhase.IN) {
           unsubscribe();
           resolve(true);
         }